asdf

+++ SVVS/src/business/controller/person/IAuthentificationController.java
@@ -9,7 +9,7 @@ package business.controller.person;
  * @author Media Markt
  */
 public interface IAuthentificationController {
-  int Authenticate(String Username, String password);
+  Long Authenticate(String Username, String password);
   
     
 }
